+++ server/db/add_buttons_to_installation.sh
@@ -10,12 +10,12 @@ if [[ $EUID > 0 ]]; then
     exit 1
 elif [[ ! -n "$3" ]]; then
     echo ""
-    echo "Usage: $0 path_to_.env_file existing_installation_name path_to_registry_csv"
+    echo "Usage: $0 path_to_.env_file existing_installation_name path_to_buttons_csv"
     echo "" 
     echo "Example: $0 ./../.env ExistingInstallation ./add_buttons.csv.example"
     echo ""
-    echo "The registry CSV file"
-    echo "MUST have the header 'button_id,unit,phone_number'"
+    echo "The buttons CSV file"
+    echo "MUST have the header 'button_id,unit,phone_number,button_serial_number'"
     echo "MUST use Unix line endings (LF), or else the phone numbers will have '\r' at the end"
     echo "MUST end with a newline, or else the last button will be silently ignored"
     echo ""
@@ -49,7 +49,7 @@ else
             echo "  Phone Number: $phone_number"
             echo "  Serial Number: $button_serial_number"
 
-        sudo PGPASSWORD=$PG_PASSWORD psql -U $PG_USER -h $PG_HOST -p $PG_PORT -d $PG_USER --set=sslmode=require -c "INSERT INTO registry (button_id, button_serial_number, unit, phone_number, installation_id) VALUES ('$button_id', '$button_serial_number', '$unit', '$phone_number', '$installation_id');"
+        sudo PGPASSWORD=$PG_PASSWORD psql -U $PG_USER -h $PG_HOST -p $PG_PORT -d $PG_USER --set=sslmode=require -c "INSERT INTO buttons (button_id, button_serial_number, unit, phone_number, installation_id) VALUES ('$button_id', '$button_serial_number', '$unit', '$phone_number', '$installation_id');"
         fi
     done < $3